absorbed BatchIndicesWithLogits
