+++ permutect/tools/prune_dataset.py
@@ -138,7 +138,7 @@ def generate_pruned_data_for_all_folds(dataset: ReadsDataset, model: PermutectMo
         summary_writer = SummaryWriter(tensorboard_dir + "/fold_" + str(pruning_fold))
         report_memory_usage(f"Pruning data from fold {pruning_fold} of {NUM_FOLDS}.")
 
-        totals_l = dataset.totals.get_marginal((BatchProperty.LABEL, )) # totals by label
+        totals_l = dataset.totals_slvra.get_marginal((BatchProperty.LABEL,)) # totals by label
         label_art_frac = totals_l[Label.ARTIFACT].item() / (totals_l[Label.ARTIFACT].item() + totals_l[Label.VARIANT].item())
         train_permutect_model(model, dataset, training_params, summary_writer=summary_writer, training_folds=[pruning_fold])
